Remove addons from the main WCProductModel
