+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/model/WCProductModel.kt
@@ -5,15 +5,12 @@ import com.google.gson.JsonArray
 import com.google.gson.JsonElement
 import com.google.gson.JsonObject
 import com.google.gson.JsonParseException
-import com.google.gson.JsonParser
 import com.yarolegovich.wellsql.core.Identifiable
 import com.yarolegovich.wellsql.core.annotation.Column
 import com.yarolegovich.wellsql.core.annotation.PrimaryKey
 import com.yarolegovich.wellsql.core.annotation.Table
 import org.wordpress.android.fluxc.model.LocalOrRemoteId.RemoteId
-import org.wordpress.android.fluxc.model.WCProductModel.AddOnsMetadataKeys.ADDONS_METADATA_KEY
 import org.wordpress.android.fluxc.model.WCProductVariationModel.ProductVariantOption
-import org.wordpress.android.fluxc.model.addons.RemoteAddonDto
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.product.CoreProductType
 import org.wordpress.android.fluxc.network.utils.getBoolean
 import org.wordpress.android.fluxc.network.utils.getLong
@@ -127,15 +124,6 @@ data class WCProductModel(@PrimaryKey @Column private var id: Int = 0) : Identif
     val attributeList: Array<ProductAttribute>
         get() = Gson().fromJson(attributes, Array<ProductAttribute>::class.java) ?: emptyArray()
 
-    val parsedMetaData: List<WCMetaData>
-        get() = runCatching { JsonParser().parse(metadata).asJsonArray }
-            .getOrNull()
-            ?.mapNotNull { element -> (element as? JsonObject)?.let { WCMetaData.fromJson(it) } }
-            ?: emptyList()
-
-    val addons: Array<RemoteAddonDto>?
-        get() = parsedMetaData.get(ADDONS_METADATA_KEY)?.addons
-
     val isConfigurable: Boolean
         get() = when (type) {
             CoreProductType.BUNDLE.value -> {
@@ -147,17 +135,6 @@ data class WCProductModel(@PrimaryKey @Column private var id: Int = 0) : Identif
             }
             else -> false
         }
-
-    @Suppress("SwallowedException", "TooGenericExceptionCaught")
-    private val WCMetaData.addons
-        get() =
-            try {
-                Gson().run {
-                    fromJson(value.jsonValue, Array<RemoteAddonDto>::class.java)
-                }
-            } catch (ex: Exception) {
-                null
-            }
 
     class ProductTriplet(val id: Long, val name: String, val slug: String) {
         fun toJson(): JsonObject {
@@ -611,7 +588,4 @@ data class WCProductModel(@PrimaryKey @Column private var id: Int = 0) : Identif
         )
     }
 
-    object AddOnsMetadataKeys {
-        const val ADDONS_METADATA_KEY = "_product_addons"
-    }
 }
